pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pypdf (>=5.8.0,<6.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "openai (>=1.97.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
